Reorganized tip strings so they can be more easily localized.
The previous structure made it difficult to provide
coherent comments for translators. It was also unnecessarily
difficult to add/remove tips.

Added comments to all localizable strings.

Also, removed two tips deemed to be overkill. This is now
(hopefully) the final set of tips.

Finally, store the current tip in SharedPreferences so that
the widget doesn't always reset to tip 0 every time Launcher
restarts (important on low-memory devices).

Bug: 2559083 (tracking the overall feature)
Change-Id: I86a29a4e8e1059e7fe2ff6792bbc4e3277951fa2